+++ apps/web/components/connected/BenefitsOfDigitalProcessesCalculator/BenefitsOfDigitalProcessesCalculator.tsx
@@ -20,32 +20,7 @@ import { formatCurrency } from '@island.is/web/utils/currency'
 
 import { formatValueForPresentation } from '../../Charts/v2/utils'
 import { t } from './translation.strings'
-
-const avinningurR = (
-  laun: number,
-  f2f: number,
-  lengd: number,
-  magn: number,
-) => {
-  return (laun / 60) * f2f * lengd * magn
-}
-
-const avinningurB = (
-  fornarkostnadur: number,
-  lengd: number,
-  f2f: number,
-  km: number,
-  kmgjald: number,
-  okuhradi: number,
-  magn: number,
-) => {
-  return (
-    (f2f * 2 * km * kmgjald +
-      (60 / okuhradi) * km * 2 * f2f * (fornarkostnadur / 60) +
-      ((f2f * lengd) / 60) * fornarkostnadur) *
-    magn
-  )
-}
+import { calculateResults, UserInput } from './utils'
 
 interface FieldProps {
   heading: string
@@ -96,31 +71,6 @@ const ResultCard = ({ title, icon, description }: ResultCardProps) => {
   )
 }
 
-interface UserInput {
-  nameOfProcess: string
-  amountPerYear: number
-  processDurationInMinutes: number
-  visitCountToCompleteProcess: number
-  averageDistanceToProcessInKilometers: number
-}
-
-interface Results {
-  /* Ávinningur stofnunar */
-  institutionGain: number
-
-  /* Ávinningur borgara */
-  citizenGain: number
-
-  /* Ígildi stöðugildis */
-  staffFreeToDoOtherThings: number
-
-  /* Eknir kílómetrar */
-  drivenKilometersSaved: number
-
-  /* Sparaðir dagar hjá fólki við að sækja sér þjónustu */
-  citizenTimeSaved: number
-}
-
 const canCalculate = (current: UserInput, previous: UserInput | null) => {
   if (
     !(
@@ -160,9 +110,7 @@ export const BenefitsOfDigitalProcessesCalculator = ({
   const { formatMessage } = useIntl()
   const { activeLocale } = useI18n()
 
-  const [resultMetadata, setResultMetadata] = useState({
-    previousInput: null as null | UserInput,
-  })
+  const [previousInput, setPreviousInput] = useState<UserInput | null>(null)
 
   const [userInput, setUserInput] = useState<UserInput>({
     amountPerYear: 0,
@@ -173,72 +121,15 @@ export const BenefitsOfDigitalProcessesCalculator = ({
     visitCountToCompleteProcess: 0,
   })
 
-  const preConditions = {
-    staffIncomePerHour:
-      slice.configJson?.['Laun starfsmanna í framþjónustu krónur á klst'] ??
-      6010,
-    citizenIncomeLossPerHour:
-      slice.configJson?.[
-        'Fórnarkostnaður borgarar (meðallaun í landi á klst)'
-      ] ?? 5122,
-    kilometerFeePerKilometer: slice.configJson?.['Km gjald pr km'] ?? 141,
-    averageDrivingSpeedInKilometersPerHour:
-      slice.configJson?.['Meðalökuhraði km/klst'] ?? 40,
-    staffHourAverageInYear:
-      slice.configJson?.['Klukkustundir í stöðugildi á ári'] ?? 1606,
-    ringRoadDistanceInKilometers:
-      slice.configJson?.['Hringvegurinn í km'] ?? 1321,
-  }
-
-  const results: Results = {
-    institutionGain: avinningurR(
-      preConditions.staffIncomePerHour,
-      userInput.visitCountToCompleteProcess,
-      userInput.processDurationInMinutes,
-      userInput.amountPerYear,
-    ),
-    citizenGain: avinningurB(
-      preConditions.citizenIncomeLossPerHour,
-      userInput.processDurationInMinutes,
-      userInput.visitCountToCompleteProcess,
-      userInput.averageDistanceToProcessInKilometers,
-      preConditions.kilometerFeePerKilometer,
-      preConditions.averageDrivingSpeedInKilometersPerHour,
-      userInput.amountPerYear,
-    ),
-    staffFreeToDoOtherThings:
-      (userInput.amountPerYear *
-        userInput.processDurationInMinutes *
-        userInput.visitCountToCompleteProcess) /
-      60 /
-      preConditions.staffHourAverageInYear,
-    drivenKilometersSaved:
-      userInput.amountPerYear *
-      userInput.visitCountToCompleteProcess *
-      2 *
-      userInput.averageDistanceToProcessInKilometers,
-    citizenTimeSaved:
-      (((userInput.visitCountToCompleteProcess *
-        2 *
-        userInput.averageDistanceToProcessInKilometers *
-        60) /
-        preConditions.averageDrivingSpeedInKilometersPerHour +
-        userInput.visitCountToCompleteProcess *
-          userInput.processDurationInMinutes) *
-        userInput.amountPerYear) /
-      60 /
-      24,
-  }
-
-  const gainPerCitizen = 0
-  const ringRoadTripsSaved =
-    results.drivenKilometersSaved / preConditions.ringRoadDistanceInKilometers
-
   const resultColumnSpan: SpanType = ['1/1', '1/2', '1/1', '1/2', '1/3']
 
   const displayResults =
-    resultMetadata.previousInput &&
-    !canCalculate(userInput, resultMetadata.previousInput)
+    previousInput && !canCalculate(userInput, previousInput)
+
+  const { results, gainPerCitizen, ringRoadTripsSaved } = calculateResults(
+    slice,
+    userInput,
+  )
 
   return (
     <Stack space={4}>
@@ -378,12 +269,10 @@ export const BenefitsOfDigitalProcessesCalculator = ({
           </Field>
 
           <Button
-            disabled={!canCalculate(userInput, resultMetadata.previousInput)}
+            disabled={!canCalculate(userInput, previousInput)}
             onClick={() => {
-              setResultMetadata({
-                previousInput: {
-                  ...userInput,
-                },
+              setPreviousInput({
+                ...userInput,
               })
             }}
           >
@@ -467,6 +356,7 @@ export const BenefitsOfDigitalProcessesCalculator = ({
                 title={formatValueForPresentation(
                   activeLocale,
                   userInput.amountPerYear,
+                  false,
                 )}
                 description={formatMessage(t.results.digitalProcessesPerYear)}
                 icon={<Icon icon="document" color="blue400" size="large" />}
